requirements.txt
fastapi
streamlit
supabase
sqlalchemy
pydantic
pandas
openpyxl
reportlab
qrcode
python-dotenv
passlib
uvicorn[standard]
python-multipart
jose
bcrypt
psycopg2-binary
st-annotated-text
streamlit-calendar
python-jose[cryptography]
python-multipart
aiofiles
pydantic[email]
